requirements.txt
python-dotenv
langchain
langgraph
langchain-community
langchain-postgres
langchain-core
langchain-huggingface
sentence-transformers
yandexcloud
python-docx
pytesseract
pillow
fastapi
uvicorn[standart]
sqlalchemy[asyncio]
asyncpg
aiofiles